requirements.txt
customtkinter
mysql-connector-python
bcrypt
pycryptodome